+++ lib/capistrano/git_copy/bundle.rb
@@ -1,5 +1,4 @@
 require 'digest/md5'
 require 'capistrano/bundler'
 require 'capistrano/git_copy/bundle/utility'
-
-load File.expand_path('../bundle/tasks/bundle.cap', __FILE__)
+require 'capistrano/git_copy/bundle/task'