+++ CHANGELOG.md
@@ -5,3 +5,4 @@
   - Package gems
   - Check for changes to prevent uploading of all gems on every deploy
   - Allow caching before actual deploy with standalone task
+  - Task to clear local and remote cache